This maybe works..

+++ src/test/lib/storage/index/partial_hash/partial_hash_index_test.cpp
@@ -568,7 +568,7 @@ TEST_F(PartialHashIndexTest, MemoryUsageNoNulls) {
   // + map size
   expected_memory_usage += sizeof(tsl::sparse_map<pmr_string, std::vector<RowID>>);
   // + number of different non-NULL values / max_load_factor * tuple size
-  expected_memory_usage += (9.0 / 0.5) * sizeof(std::pair<pmr_string, std::vector<RowID>>);
+  expected_memory_usage += static_cast<size_t>((9.0 / 0.5) * sizeof(std::pair<pmr_string, std::vector<RowID>>));
   // + number of different non-NULL values * vector size
   expected_memory_usage += 9 * sizeof(std::vector<RowID>);
   // + number of non-NULL values * RowID
@@ -610,7 +610,7 @@ TEST_F(PartialHashIndexTest, MemoryUsageNulls) {
   // + map size
   expected_memory_usage += sizeof(tsl::sparse_map<pmr_string, std::vector<RowID>>);
   // + number of different non-NULL values / max_load_factor * tuple size
-  expected_memory_usage += (0.0 / 0.5) * sizeof(std::pair<pmr_string, std::vector<RowID>>);
+  expected_memory_usage += static_cast<size_t>((0.0 / 0.5) * sizeof(std::pair<pmr_string, std::vector<RowID>>));
   // + number of different non-NULL values * vector size
   expected_memory_usage += 0 * sizeof(std::vector<RowID>);
   // + number of non-NULL values * RowID
@@ -653,7 +653,7 @@ TEST_F(PartialHashIndexTest, MemoryUsageMixed) {
   // + map size
   expected_memory_usage += sizeof(tsl::sparse_map<pmr_string, std::vector<RowID>>);
   // + number of different non-NULL values / max_load_factor * tuple size
-  expected_memory_usage += (9.0/ 0.5) * sizeof(std::pair<pmr_string, std::vector<RowID>>);
+  expected_memory_usage += static_cast<size_t>((9.0/ 0.5) * sizeof(std::pair<pmr_string, std::vector<RowID>>));
   // + number of different non-NULL values * vector size
   expected_memory_usage += 9 * sizeof(std::vector<RowID>);
   // + number of non-NULL values * RowID
@@ -694,7 +694,7 @@ TEST_F(PartialHashIndexTest, MemoryUsageEmpty) {
   // + map size
   expected_memory_usage += sizeof(tsl::sparse_map<pmr_string, std::vector<RowID>>);
   // + number of different non-NULL values / max_load_factor * tuple size
-  expected_memory_usage += (0.0 / 0.5) * sizeof(pmr_string);
+  expected_memory_usage += static_cast<size_t>((0.0 / 0.5) * sizeof(pmr_string));
   // + number of different non-NULL values * vector size
   expected_memory_usage += 0 * sizeof(std::vector<RowID>);
   // + number of non-NULL values * RowID
@@ -736,7 +736,7 @@ TEST_F(PartialHashIndexTest, MemoryUsageNoChunk) {
   // + map size
   expected_memory_usage += sizeof(tsl::sparse_map<pmr_string, std::vector<RowID>>);
   // + number of different non-NULL values / max_load_factor * tuple size
-  expected_memory_usage += (0.0 / 0.5) * sizeof(std::pair<pmr_string, std::vector<RowID>>);
+  expected_memory_usage += static_cast<size_t>((0.0 / 0.5) * sizeof(std::pair<pmr_string, std::vector<RowID>>));
   // + number of different non-NULL values * vector size
   expected_memory_usage += 0 * sizeof(std::vector<RowID>);
   // + number of non-NULL values * RowID
